Changes on CartContorller

+++ backend/WebApi/Controllers/Orders/CartController.cs
@@ -1,8 +1,10 @@
+using System.Security.Claims;
 using Business.Abstract.Orders;
 using Entities.Concrete.Orders;
 using Entities.Concrete.Users;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.FileSystemGlobbing.Internal.PathSegments;
 
 namespace WebApi.Controllers.Orders
 {
@@ -19,6 +21,7 @@ namespace WebApi.Controllers.Orders
         }
 
         [HttpGet("getall")]
+        [Authorize(Roles = Roles.Admin)]
         public IActionResult GetAll()
         {
             var result = _cartService.GetAll();
@@ -28,6 +31,7 @@ namespace WebApi.Controllers.Orders
             return BadRequest(result);
         }
 
+        [Authorize(Roles = Roles.Admin)]
         [HttpGet("getbyid/{id}")]
         public IActionResult GetById(int id)
         {
@@ -38,14 +42,12 @@ namespace WebApi.Controllers.Orders
             return NotFound(result);
         }
 
-        [HttpGet("getbyuserid/{userId}")]
-        public IActionResult GetCartByUserId(int userId)
+        [HttpGet("getbyuser")]
+        public IActionResult GetCartByUserId()
         {
-            // Add user verification
-            var userIdClaim = User.FindFirst("nameid")?.Value;
-            if (userIdClaim != userId.ToString())
-            {
-                // return Forbid();
+            var userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+            if (userId == null) {
+                return BadRequest("You need to be logged in.");
             }
 
             var result = _cartService.GetCartByUserId(userId);
@@ -55,6 +57,7 @@ namespace WebApi.Controllers.Orders
             return NotFound(result);
         }
 
+        [Authorize(Roles = Roles.Admin)]
         [HttpPost("add")]
         public IActionResult Add([FromBody] Cart cart)
         {
@@ -65,6 +68,7 @@ namespace WebApi.Controllers.Orders
             return BadRequest(result);
         }
 
+        [Authorize(Roles = Roles.Admin)]
         [HttpPut("update/{id}")]
         public IActionResult Update(int id, [FromBody] Cart cart)
         {
@@ -77,7 +81,8 @@ namespace WebApi.Controllers.Orders
 
             return BadRequest(result);
         }
-
+        
+        [Authorize(Roles = Roles.Admin)]
         [HttpDelete("delete/{id}")]
         public IActionResult Delete(int id)
         {
@@ -93,8 +98,12 @@ namespace WebApi.Controllers.Orders
         }
 
         [HttpPost("additem")]
-        public IActionResult AddItemToCart([FromQuery] int userId, [FromQuery] int productId, [FromQuery] int quantity)
+        public IActionResult AddItemToCart([FromQuery] int productId, [FromQuery] int quantity)
         {
+            var userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+            if (userId == null) {
+                return BadRequest("You need to be logged in.");
+            }
             var result = _cartService.AddItemToCart(userId, productId, quantity);
             if (result.Success)
                 return Ok(result);
@@ -102,8 +111,12 @@ namespace WebApi.Controllers.Orders
         }
 
         [HttpDelete("removeitem")]
-        public IActionResult RemoveItemFromCart([FromQuery] int userId, [FromQuery] int productId)
+        public IActionResult RemoveItemFromCart([FromQuery] int productId)
         {
+            var userId = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+            if (userId == null) {
+                return BadRequest("You need to be logged in.");
+            }
             var result = _cartService.RemoveItemFromCart(userId, productId);
             if (result.Success)
                 return Ok(result);
